run.sh can now be run locally - i.e. full system

+++ run.sh
@@ -13,11 +13,19 @@ create)
     ;;
 esac
 
-./expand_variables.sh
-python ./awslogs-agent-setup.py -n -r eu-west-1 -c ./awslogs.conf
-ps -eaf | grep awslogs | grep -v grep | awk -F' ' '{print $2'} | xargs kill -9
-supervisord -c /etc/supervisor.conf &
+# remove if statement after migration to Azure is complete.
+# python ./awslogs-agent-setup.py -n -r eu-west-1 -c ./awslogs.conf
+if [[ $CLOUD_PROVIDER == "azure" || $DISABLE_CLOUDWATCH == "true" ]]; then
+# ps -eaf | grep awslogs | grep -v grep | awk -F' ' '{print $2'} | xargs kill -9
+    echo "Running on Azure"
+# supervisord -c /etc/supervisor.conf &
+else
+    ./expand_variables.sh
+    python ./awslogs-agent-setup.py -n -r eu-west-1 -c ./awslogs.conf
+    ps -eaf | grep awslogs | grep -v grep | awk -F' ' '{print $2'} | xargs kill -9
+    supervisord -c /etc/supervisor.conf &
+fi
 
 echo "Running app"
 
-bundle exec unicorn -p 8080 -c ./config/unicorn.rb -E production
+bundle exec unicorn -p ${PORT:-8080} -c ./config/unicorn.rb -E ${RAILS_ENV:-production}
